+++ docs/changelog.rst
@@ -6,6 +6,21 @@ Changelog
 Detailed version changes.
 You can also view the changelog of the original library, fortnitepy, `here <https://fortnitepy.readthedocs.io/en/latest/changelog.html>`_.
 
+v1.0.2
+------
+
+Thanks to `Charlesbug <https://github.com/charlesbug>`_ for adding most of this update.
+
+Added
+~~~~~
+
+- Added :meth:`ClientPartyMember.set_sidekick_emote()` & :meth:`ClientPartyMember.clear_sidekick_emote()`
+
+Removed
+~~~~~~~
+
+- Removed :meth:`ClientPartyMember.hifive_sidekick()` & :meth:`ClientPartyMember.clear_hifive()`, use `set_sidekick_emote("Interact")` now to hi-five your sidekick.
+
 v1.0.1
 ------